add (failing) import tests

+++ test/prog/run_all.sh
@@ -3,6 +3,9 @@
 set -ex
 
 for xl_file in test/prog/*.xl; do
+        if grep skip-run $xl_file; then
+                continue
+        fi
         dist/expelc $xl_file a.out
         dist/runexpel a.out
         rm a.out
